Move helper methods to the new task class
